intern_projects/AI-Driven-Renewable-Energy-Optimization/requirements.txt
torch
torchvision
scikit-learn

streamlit
Pillow
numpy
pandas

# Add these new ones for efficiency prediction
plotly>=5.0.0
seaborn>=0.11.0
matplotlib>=3.5.0